+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `frame-inclinometer` would show `pitch (old)` to indicate the old incorrect calculation, but that actually implies to `roll` and not `pitch`
 - Sps kitchen sink did not recognize inclinometer calibration values as valid entries
 
 ## [1.21.1]
