+++ src/app.rs
@@ -1,10 +1,7 @@
-use std::{
-    sync::mpsc::{self, Receiver, Sender},
-    time::Duration,
-};
+use std::time::Duration;
 
 use crate::{
-    app::download::{ENDPOINT_DOWNLOAD_LATEST, ENDPOINT_DOWNLOAD_TODAY},
+    app::download::{DownloadManager, DownloadMessage},
     plot::LogPlotUi,
     util::format_data_size,
 };
@@ -81,8 +78,8 @@ impl Default for App {
             font_size: Self::DEFAULT_FONT_SIZE,
             font_size_init: false,
             error_message: None,
-            download_host: "localhost".to_string(),
-            download_port: "8080".to_string(),
+            download_host: "192.168.1.60".to_string(),
+            download_port: "9091".to_string(),
 
             #[cfg(all(not(target_arch = "wasm32"), feature = "mqtt"))]
             mqtt: crate::mqtt::Mqtt::default(),
@@ -133,30 +130,7 @@ impl eframe::App for App {
     /// Called each time the UI needs repainting, which may be many times per second.
     fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
         self.poll_picked_files();
-        // Poll download messages
-        for msg in self.download_manager.poll() {
-            match msg {
-                DownloadMessage::Success(filename) => {
-                    self.toasts
-                        .success(format!("Downloaded: {}", filename))
-                        .duration(Some(Duration::from_secs(5)));
-                }
-                DownloadMessage::Error(err) => {
-                    self.toasts
-                        .error(format!("Download failed: {}", err))
-                        .duration(Some(Duration::from_secs(10)));
-                }
-                DownloadMessage::Progress {
-                    downloaded_bytes,
-                    total_bytes,
-                } => {
-                    self.download_manager
-                        .update_progress(downloaded_bytes, total_bytes);
-                }
-                DownloadMessage::Finished => {}
-            }
-            ctx.request_repaint();
-        }
+        self.poll_download_messages(ctx);
 
         if !self.font_size_init {
             configure_text_styles(ctx, self.font_size);
@@ -191,7 +165,7 @@ impl eframe::App for App {
                 egui::warn_if_debug_build(ui);
             });
         });
-        show_download_window(self, ctx);
+        download::show_download_window(self, ctx);
 
         self.toasts.show(ctx);
     }
@@ -254,6 +228,32 @@ impl App {
             Ok(Some(new_plot_ui_state)) => self.load_new_plot_ui_state(new_plot_ui_state),
             Err(e) => self.error_message = Some(e.to_string()),
             Ok(None) => (),
+        }
+    }
+
+    fn poll_download_messages(&mut self, ctx: &egui::Context) {
+        for msg in self.download_manager.poll() {
+            match msg {
+                DownloadMessage::Success { filename } => {
+                    self.toasts
+                        .success(format!("Downloaded: {filename}"))
+                        .duration(Some(Duration::from_secs(5)));
+                }
+                DownloadMessage::Error(err) => {
+                    self.toasts
+                        .error(format!("Download failed: {err}"))
+                        .duration(Some(Duration::from_secs(10)));
+                }
+                DownloadMessage::Progress {
+                    downloaded_bytes,
+                    total_bytes,
+                } => {
+                    self.download_manager
+                        .update_progress(downloaded_bytes, total_bytes);
+                }
+                DownloadMessage::Finished => {}
+            }
+            ctx.request_repaint();
         }
     }
 }
@@ -476,132 +476,4 @@ fn show_theme_toggle_buttons(ui: &mut egui::Ui) {
     });
 
     ui.ctx().set_theme(theme_preference);
-}
-
-pub(crate) enum DownloadMessage {
-    Success(String), // filename
-    Error(String),   // error message
-    Progress {
-        downloaded_bytes: u64,
-        total_bytes: u64,
-    },
-    Finished,
-}
-
-fn show_download_window(app: &mut App, ctx: &egui::Context) {
-    if !app.show_download_window {
-        return;
-    }
-    if app.download_manager.in_progress() {
-        ctx.request_repaint_after(Duration::from_millis(50));
-    }
-
-    egui::Window::new("Download Logs")
-        .collapsible(false)
-        .resizable(true)
-        .open(&mut app.show_download_window)
-        .show(ctx, |ui| {
-            ui.add_enabled_ui(!app.download_manager.in_progress(), |ui| {
-                ui.horizontal(|ui| {
-                    ui.label("Host:");
-                    ui.text_edit_singleline(&mut app.download_host);
-                    ui.label("Port:");
-                    ui.add_sized(
-                        [80.0, 24.0],
-                        egui::TextEdit::singleline(&mut app.download_port),
-                    );
-                });
-            });
-
-            ui.separator();
-
-            if app.download_manager.in_progress() {
-                ui.vertical_centered(|ui| {
-                    ui.add(egui::ProgressBar::new(app.download_manager.progress).show_percentage());
-                    ui.label(&app.download_manager.status_text);
-                });
-            } else if ui.button("Download Latest data").clicked() {
-                app.download_manager.start_download(
-                    app.download_host.clone(),
-                    app.download_port.clone(),
-                    ENDPOINT_DOWNLOAD_LATEST.to_owned(),
-                );
-            } else if ui.button("Download Today's Data").clicked() {
-                app.download_manager.start_download(
-                    app.download_host.clone(),
-                    app.download_port.clone(),
-                    ENDPOINT_DOWNLOAD_TODAY.to_owned(),
-                );
-            }
-        });
-}
-
-pub struct DownloadManager {
-    tx: Sender<DownloadMessage>,
-    rx: Receiver<DownloadMessage>,
-    in_progress: bool,
-    progress: f32,
-    status_text: String,
-}
-
-impl DownloadManager {
-    pub fn new() -> Self {
-        let (tx, rx) = mpsc::channel();
-        Self {
-            tx,
-            rx,
-            in_progress: false,
-            progress: 0.0,
-            status_text: String::new(),
-        }
-    }
-
-    pub fn start_download(&mut self, host: String, port: String, endpoint: String) {
-        if self.in_progress {
-            return;
-        }
-        self.in_progress = true;
-        self.progress = 0.0;
-        self.status_text = "Connecting...".to_string();
-
-        let tx = self.tx.clone();
-        std::thread::spawn(move || {
-            let result = download::download_zip(&host, &port, tx.clone(), &endpoint);
-            match result {
-                Ok(filename) => {
-                    let _ = tx.send(DownloadMessage::Success(filename));
-                }
-                Err(e) => {
-                    let _ = tx.send(DownloadMessage::Error(e.to_string()));
-                }
-            }
-            let _ = tx.send(DownloadMessage::Finished);
-        });
-    }
-
-    pub(crate) fn poll(&mut self) -> Vec<DownloadMessage> {
-        let mut messages = Vec::new();
-        while let Ok(msg) = self.rx.try_recv() {
-            if matches!(msg, DownloadMessage::Finished) {
-                self.in_progress = false;
-            }
-            messages.push(msg);
-        }
-        messages
-    }
-
-    pub fn in_progress(&self) -> bool {
-        self.in_progress
-    }
-
-    pub fn update_progress(&mut self, downloaded: u64, total: u64) {
-        if total > 0 {
-            self.progress = downloaded as f32 / total as f32;
-        }
-        self.status_text = format!(
-            "{} / {}",
-            format_data_size(downloaded as usize),
-            format_data_size(total as usize)
-        );
-    }
 }
