+++ src/app/download.rs
@@ -1,97 +1,149 @@
-use std::fs::{self, File};
-use std::io::{Read, Write};
-use std::path::PathBuf;
-use std::sync::mpsc::Sender;
-use std::time::Duration;
+use std::{
+    sync::mpsc::{self, Receiver, Sender},
+    thread,
+    time::Duration,
+};
 
-use anyhow::bail;
+use crate::{
+    App,
+    app::download::{
+        self,
+        connection::{ENDPOINT_DOWNLOAD_LATEST, ENDPOINT_DOWNLOAD_TODAY},
+    },
+    util::format_data_size,
+};
 
-use crate::app::DownloadMessage;
+mod connection;
 
-pub(crate) const ENDPOINT_DOWNLOAD_LATEST: &str = "/api/download/latest";
-pub(crate) const ENDPOINT_DOWNLOAD_TODAY: &str = "/api/download/today";
-
-/// Downloads a zip file from the specified host and port.
-/// Returns the filename of the downloaded file on success.
-pub fn download_zip(
-    host: &str,
-    port: &str,
-    tx: Sender<DownloadMessage>,
-    endpoint: &str,
-) -> anyhow::Result<String> {
-    // Construct the URL
-    let url = format!("http://{host}:{port}{endpoint}");
-
-    // Use a client with very long timeouts for zip generation
-    let client = reqwest::blocking::Client::builder()
-        .timeout(Duration::from_secs(30 * 60)) // 30 minutes total timeout
-        .connect_timeout(Duration::from_secs(30)) // 30 seconds to establish connection
-        .build()?;
-
-    let mut response = client.get(&url).send()?;
-
-    // Check if the response is successful
-    if !response.status().is_success() {
-        bail!("Server returned status: {}", response.status());
+pub(super) fn show_download_window(app: &mut App, ctx: &egui::Context) {
+    if !app.show_download_window {
+        return;
+    }
+    if app.download_manager.in_progress() {
+        ctx.request_repaint_after(Duration::from_millis(50));
     }
 
-    // Get total size from Content-Length header for the progress bar
-    let total_size = response
-        .headers()
-        .get(reqwest::header::CONTENT_LENGTH)
-        .and_then(|v| v.to_str().ok())
-        .and_then(|s| s.parse::<u64>().ok())
-        .unwrap_or(0);
+    egui::Window::new("Download Logs")
+        .collapsible(false)
+        .resizable(true)
+        .open(&mut app.show_download_window)
+        .show(ctx, |ui| {
+            ui.add_enabled_ui(!app.download_manager.in_progress(), |ui| {
+                ui.horizontal(|ui| {
+                    ui.label("Host:");
+                    ui.text_edit_singleline(&mut app.download_host);
+                    ui.label("Port:");
+                    ui.add_sized(
+                        [80.0, 24.0],
+                        egui::TextEdit::singleline(&mut app.download_port),
+                    );
+                });
+            });
 
-    log::info!("Receiving {total_size} bytes");
+            ui.separator();
 
-    // Create the plotinator directory in the downloads folder
-    let download_dir = get_download_directory();
-    let plotinator_dir = download_dir.join("plotinator");
-    fs::create_dir_all(&plotinator_dir)?;
-
-    // Generate a filename with timestamp
-    let timestamp = chrono::Utc::now().format("%Y%m%d_%H%M%S");
-    let filename = format!("plotinator_download_{}.zip", timestamp);
-    let file_path = plotinator_dir.join(&filename);
-
-    // Create the file and prepare for chunked writing
-    let mut file = File::create(&file_path)?;
-    let mut downloaded: u64 = 0;
-    let mut buffer = [0; 8192]; // 8KB buffer
-
-    // Read the response in chunks
-    while let Ok(bytes_read) = response.read(&mut buffer) {
-        log::trace!("Read chunk: {bytes_read} bytes");
-        if bytes_read == 0 {
-            break; // End of stream
-        }
-
-        file.write_all(&buffer[..bytes_read])?;
-        downloaded += bytes_read as u64;
-
-        // Send progress update to the UI thread
-        tx.send(DownloadMessage::Progress {
-            downloaded_bytes: downloaded,
-            total_bytes: total_size,
-        })?;
-    }
-
-    file.flush()?;
-
-    Ok(filename)
+            if app.download_manager.in_progress() {
+                ui.vertical_centered(|ui| {
+                    ui.add(egui::ProgressBar::new(app.download_manager.progress).show_percentage());
+                    ui.label(&app.download_manager.status_text);
+                });
+            } else if ui.button("Download Latest data").clicked() {
+                app.download_manager.start_download(
+                    app.download_host.clone(),
+                    app.download_port.clone(),
+                    ENDPOINT_DOWNLOAD_LATEST.to_owned(),
+                );
+            } else if ui.button("Download Today's Data").clicked() {
+                app.download_manager.start_download(
+                    app.download_host.clone(),
+                    app.download_port.clone(),
+                    ENDPOINT_DOWNLOAD_TODAY.to_owned(),
+                );
+            }
+        });
 }
 
-fn get_download_directory() -> PathBuf {
-    dirs::download_dir()
-        .or_else(|| dirs::home_dir().map(|home| home.join("Downloads")))
-        .unwrap_or_else(|| {
-            let fallback = std::env::current_dir()
-                .unwrap_or_else(|_| PathBuf::from("."))
-                .join("downloads");
+pub(crate) enum DownloadMessage {
+    Success {
+        filename: String,
+    },
+    Error(String),
+    Progress {
+        downloaded_bytes: u64,
+        total_bytes: u64,
+    },
+    Finished,
+}
 
-            // Try to create the directory
-            let _ = fs::create_dir_all(&fallback);
-            fallback
-        })
+pub struct DownloadManager {
+    tx: Sender<DownloadMessage>,
+    rx: Receiver<DownloadMessage>,
+    in_progress: bool,
+    progress: f32,
+    status_text: String,
+}
+
+impl DownloadManager {
+    pub fn new() -> Self {
+        let (tx, rx) = mpsc::channel();
+        Self {
+            tx,
+            rx,
+            in_progress: false,
+            progress: 0.0,
+            status_text: String::new(),
+        }
+    }
+
+    pub fn start_download(&mut self, host: String, port: String, endpoint: String) {
+        if self.in_progress {
+            return;
+        }
+        self.in_progress = true;
+        self.progress = 0.0;
+        self.status_text = "Connecting...".to_string();
+
+        let tx = self.tx.clone();
+        thread::Builder::new()
+            .name("downloader".into())
+            .spawn(move || {
+                let result = connection::download_zip(&host, &port, tx.clone(), &endpoint);
+                match result {
+                    Ok(filename) => {
+                        let _ = tx.send(DownloadMessage::Success { filename });
+                    }
+                    Err(e) => {
+                        let _ = tx.send(DownloadMessage::Error(e.to_string()));
+                    }
+                }
+                let _ = tx.send(DownloadMessage::Finished);
+            })
+            .expect("Failed spawning download thread");
+    }
+
+    pub(crate) fn poll(&mut self) -> Vec<DownloadMessage> {
+        let mut messages = Vec::new();
+        while let Ok(msg) = self.rx.try_recv() {
+            if matches!(msg, DownloadMessage::Finished) {
+                self.in_progress = false;
+            }
+            messages.push(msg);
+        }
+        messages
+    }
+
+    pub fn in_progress(&self) -> bool {
+        self.in_progress
+    }
+
+    pub fn update_progress(&mut self, downloaded: u64, total: u64) {
+        if total > 0 {
+            self.progress = downloaded as f32 / total as f32;
+        }
+        self.status_text = format!(
+            "{} / {}",
+            format_data_size(downloaded as usize),
+            format_data_size(total as usize)
+        );
+    }
 }
